Service layer now has Interfaces
