+++ src/main/java/io/dashboardhub/pipelinedashboard/controller/ProjectController.java
@@ -1,8 +1,7 @@
 package io.dashboardhub.pipelinedashboard.controller;
 
 import io.dashboardhub.pipelinedashboard.domain.Project;
-import io.dashboardhub.pipelinedashboard.service.ProjectService;
-import io.dashboardhub.pipelinedashboard.service.UserService;
+import io.dashboardhub.pipelinedashboard.service.ProjectServiceImpl;
 import io.dashboardhub.pipelinedashboard.service.exception.PermissionDeniedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -19,11 +18,11 @@ import javax.validation.Valid;
 public class ProjectController {
 
     @Autowired
-    private ProjectService projectService;
+    private ProjectServiceImpl projectServiceImpl;
 
     @RequestMapping(value = {"/project"}, method = RequestMethod.GET)
     public String list(Model model) {
-        model.addAttribute("projects", projectService.findAllByCurrentUser());
+        model.addAttribute("projects", projectServiceImpl.findAllByCurrentUser());
 
         return "project/list";
     }
@@ -41,7 +40,7 @@ public class ProjectController {
     @RequestMapping(value = {"/project/edit/{uid}"}, method = RequestMethod.GET)
     public String edit(@PathVariable("uid") String uid, Model model, final RedirectAttributes redirectAttributes) {
 
-        Project project = projectService.findByUid(uid);
+        Project project = projectServiceImpl.findByUid(uid);
         if (project == null) {
             redirectAttributes.addFlashAttribute("error", "Not found");
 
@@ -68,7 +67,7 @@ public class ProjectController {
             return "project/add";
         }
 
-        projectService.save(project);
+        projectServiceImpl.save(project);
         redirectAttributes.addFlashAttribute("success", "Successfully saved");
 
         return "redirect:/project";
@@ -89,7 +88,7 @@ public class ProjectController {
         }
 
         try {
-            projectService.save(project);
+            projectServiceImpl.save(project);
             redirectAttributes.addFlashAttribute("success", "Successfully saved");
         } catch (PermissionDeniedException e) {
             redirectAttributes.addFlashAttribute("error", "Permission denied");
@@ -101,7 +100,7 @@ public class ProjectController {
 
     @RequestMapping(value = {"/project/delete/{uid}"}, method = RequestMethod.GET)
     public String delete(@PathVariable("uid") String uid, final RedirectAttributes redirectAttributes, Model model) {
-        Project project = projectService.findByUid(uid);
+        Project project = projectServiceImpl.findByUid(uid);
         if (project == null) {
             redirectAttributes.addFlashAttribute("error", "Not found");
 
@@ -109,7 +108,7 @@ public class ProjectController {
         }
 
         try {
-            projectService.delete(project);
+            projectServiceImpl.delete(project);
             redirectAttributes.addFlashAttribute("success", "Successfully removed");
         } catch (PermissionDeniedException e) {
             redirectAttributes.addFlashAttribute("error", "Permission denied");
