+++ src/main/java/io/dashboardhub/pipelinedashboard/listener/AuthenticationListener.java
@@ -1,7 +1,7 @@
 package io.dashboardhub.pipelinedashboard.listener;
 
 import io.dashboardhub.pipelinedashboard.domain.User;
-import io.dashboardhub.pipelinedashboard.service.UserService;
+import io.dashboardhub.pipelinedashboard.service.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.security.authentication.event.InteractiveAuthenticationSuccessEvent;
@@ -13,18 +13,18 @@ import java.util.Date;
 public class AuthenticationListener implements ApplicationListener<InteractiveAuthenticationSuccessEvent> {
 
     @Autowired
-    private UserService userService;
+    private UserServiceImpl userServiceImpl;
 
     @Override
     public void onApplicationEvent(final InteractiveAuthenticationSuccessEvent event) {
-        User user = userService.findByUsername(event.getAuthentication().getName());
+        User user = userServiceImpl.findByUsername(event.getAuthentication().getName());
 
         if (user == null) {
-            userService.save(new User(event.getAuthentication().getName()));
+            userServiceImpl.save(new User(event.getAuthentication().getName()));
         }
 
-        user = userService.findByUsername(event.getAuthentication().getName());
+        user = userServiceImpl.findByUsername(event.getAuthentication().getName());
         user.setLastLoggedIn(new Date().toString());
-        userService.save(user);
+        userServiceImpl.save(user);
     }
 }
